fixup! refactor: isolate host audio probe backend
